requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
python-dotenv==0.19.0
openai==0.28
supabase
